requirements.txt
cffi==1.14.3
cryptography==3.1.1
fastapi==0.61.1
pycparser==2.20
pydantic==1.6.1
PyMySQL==0.10.1
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.5
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
starlette==0.13.6
